requirements.txt
Flask
PyMySQL
Flask-SQLAlchemy
cryptography
psycopg2-binary
python-dateutil
